Hotfix: Merging difficulties fixed
Accidentally built another folder when merging
