+++ EpicColors/EpicColors.csproj
@@ -14,11 +14,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.10.0" />
     <PackageReference Include="Reactor.MSBuild" Version="*" PrivateAssets="all" />
   </ItemGroup>
   
   <Target Name="Copy" AfterTargets="Build">
-    <Copy SourceFiles="$(OutputPath)$(AssemblyName).dll" DestinationFolder="$(AmongUs)/BepInEx/plugins/" Condition="'$(AmongUs)' != ''"/>
+    <Copy SourceFiles="$(OutputPath)$(AssemblyName).dll" DestinationFolder="$(AmongUs)/BepInEx/plugins/" Condition="'$(AmongUs)' != ''" />
     <Copy SourceFiles="$(OutputPath)$(AssemblyName).dll" DestinationFiles="../Build/$(AssemblyName).dll" />
   </Target>
 </Project>
