+++ EpicColors/Handler/ConfigBuilder.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using System.Text;
+using UnityEngine;
 
 namespace EpicColors
 {
