+++ EpicColors/Patches/Animated/Animated.cs
