+++ EpicColors/Patches/Animated/AnimatedColors.cs
