+++ EpicColors/Patches/ColorTypes/BaseColor.cs
