+++ EpicColors/Patches/ColorTypes/Refresh.cs
