+++ EpicColors/Patches/ColorTypes/Static.cs
