+++ EpicColors/Patches/EpicComponent.cs
