+++ EpicColors/Patches/Watermark.cs
@@ -1,9 +1,5 @@
 using HarmonyLib;
-using System;
-using System.Linq;
-using System.IO;
 using UnityEngine;
-using static EpicColors.ConverterHelper;
 using static EpicColors.CustomColorHandler;
 
 namespace EpicColors
@@ -33,31 +29,12 @@ namespace EpicColors
                 }
 
                 if (AmongUsClient.Instance.GameState != InnerNet.InnerNetClient.GameStates.Started) {
-                    t.text += "\nEpicColors by Devs-Us\n";
+                    Debug.logger.Log(Author);
+                    t.text += "\nEpicColors by Devs-Us <size=80%>v1.0.0</size>\n";
                     t.text += $"You are using {id.GetColorName()}\n";
                     t.text += Author;
                 }
             }
-        }
-
-        public static string ToAuthor(this int colorId) {
-            var name = "";
-            foreach (var author in TxtContentList)
-                if (author.StartsWith("author;") && 
-                IsUsingCustomColor(colorId, out bool customColor) && customColor) {
-                    var finalAuthor = author.Replace("author;","");
-                    name += finalAuthor + "\n";
-                }
-            return name;
-        }
-
-        public static string GetColorName(this int colorId) {
-            var name = IncludeBuiltinColor() ? AllCCList[colorId-OldPaletteCount].RealColorName() 
-            : CustomColorList[colorId-OldPaletteCount].RealColorName();
-            var color = Palette.PlayerColors[colorId].ToHexString();
-
-            return IsUsingCustomColor(colorId+1, out _) 
-            ? $"You are using <color=#{color}>{name}</color>\n" : "";
         }
     }
 }
